Add: getting monday and sunday
